Delay evaluation of completion conditions as much as possible
This patch causes fish to first determine to the best of its ability whether or
not a potential completion could/would be used before actually checking if its
`-n` conditions have been met.

Determining whether or not a completion will be used is entirely in compiled
code and involves a couple of character and string comparisons whereas
determining if a completion's condition is met requires the (relatively heavy)
execution of a subshell in addition to the cost incurred by all the forking and
executions the shell script completion conditions would make.

For the simple case of `complete -C"git diff -c"`, previously fish would
evaluate 1459 condition *lists* (not even individual completions), each of which
might trigger the invocation of a subshell and a job (if the exact same
condition string wasn't already calculated and cached for the same overall
completion attempt).

This revised version cuts down the number of evaluated completion conditions for
that same sample completion prompt to 312.

Improvements scale with the number of installed completions for a given command,
the number of conditions attached to said completions, and the actual condition
predicates themselves. The actual wins are more modest than I'd hoped for the
average case - although the sky is the limit if these changes avoid one or more
pathological completion conditions; benchmarking indicates that `fish -c
"complete -C'git ...'"` now executes 1-7% faster on a variety of different `git`
completions (with an error margin of ±2%).