function: Add a bunch of awkward helper functions
I have no hope of getting function_props_ref_t through the FFI,
so instead we give a bunch of functions.

Note: These are hopelessly unsafe - if the function doesn't exist,
they'll return dummy data.

Also instead of a filename_ref_t we return a wcstring, which is just
empty if the filename is NULL - under the assumption that a filename
can never be empty (if it exists).