Use wide strings with path_get_paths_from_parser()
This eliminates converting wide-to-narrow in C++ then narrow-back-to-wide in
rust.

Though VectorElement is implemented for CxxWString, autocxx reads the
CxxVector<CxxWString> as CxxVector<std_basic_string_wchar_t_AutocxxConcrete> so
I had to use a transmute to cast it the reference to the correct type. This
needs further looking into.

Also, the printf-compat crate/fork doesn't implement ToArg for CxxWString, so we
have to first get a wstr from the CxxWString and pass that to to the printf
macro.