Synchronize globals between C++ and Rust
When we were calling the rust version of the global init functions in
fallback.cpp/fallback.rs and common.cpp/common.rs, they were getting set in one
but not the other. Use `extern` definitions in C++ for some to read the rust
values and for others just have rust also call the C++ version, too.

Without this, tests fail and fish segfaults when it tries to read the
uninitialized or stale global variables like g_fish_emoji_width or
g_ellipsis_str;