Get rid of locking in curses fast-path
The global `curses::Term` singleton is no longer behind a mutex and uses
interior mutability to store overrides. There is no locking if no terminal
sequences have been overwritten, and any locking if they /have/ been overwritten
is only momentary and released before `Term::get()` or `Term::set()` return to
the caller.

This necessitated storing the overwrite strings as Arc<String> instead of String
because we need to return a borrow after the mutex lock expires, which isn't
normally possible. A new SuperCow<T> has been added that supports the regular
Cow's Owned/Borrowed plus a custom Shared variant that is used when holding onto
a copy of an Arc<T>. This isn't just for the curses module; I want to also use
this when returning `Environment` variables.